README.md
What is SP-TBI Application?
SP-TBI Application is an Android based application used for ease of communication between the incubators, pre-incubators and the Technology Business Incubator. 
We have the incubators and pre-incubators filling in their information on the application. With the help of this, SP-TBI can not only make communication easier but also have a constant updated database of the start-ups joining and leaving. 

IDE used:

1. Android studio

Data structure implemented:

1. HashMap - to retrieve data from Firebase database 
2. ArrayList -  to store the list of users
3. Adapter - to store the list of users and display it to the admin
4. Recycler view - to display the incubators
5. Bitmap - to store the compressed form of a image

Back end:

1. Firebase - for real-time database, media storage and user authentication

API used:

1. Picasso 
2. Firebase service
3. Google service
4. Image cropper 
5. Image compressor